+++ Test/TestHelpers/StubFileStoreCacheClass.cs
@@ -4,9 +4,7 @@
 using System.Text.Encodings.Web;
 using System.Text.Json;
 using Microsoft.Extensions.Caching.Distributed;
-using Microsoft.Extensions.Primitives;
 using Net.DistributedFileStoreCache;
-using Net.DistributedFileStoreCache.SupportCode;
 
 namespace Test.TestHelpers;
 
@@ -68,9 +66,36 @@ public class StubFileStoreCacheClass : StubFileStoreCacheString, IDistributedFil
     /// <param name="token">Optional. The <see cref="T:System.Threading.CancellationToken" /> used to propagate notifications that the operation should be canceled.</param>
     /// <typeparam name="T">A class which can be created</typeparam>
     public async Task SetClassAsync<T>(string key, T yourClass, DistributedCacheEntryOptions? options = null,
-        CancellationToken token = new CancellationToken()) where T : class, new()
+        CancellationToken token = new ()) where T : class, new()
     {
         var jsonString = JsonSerializer.Serialize(yourClass, _jsonOptions);
         await SetAsync(key, jsonString, options, token);
+    }
+
+    /// <summary>Serializes all the values in each KeyValue using the T type and save each into the cache</summary>
+    /// <param name="manyEntries">List of KeyValuePairs to be added to the cache, with the values being serialized.</param>
+    /// <param name="options">Optional: The cache options for the value.</param>
+    /// <typeparam name="T">A class which contains the data to stored as JSON in the cache</typeparam>
+    public void SetManyClass<T>(List<KeyValuePair<string, T>> manyEntries, DistributedCacheEntryOptions? options) where T : class, new()
+    {
+        foreach (var keyValue in manyEntries)
+        {
+            var jsonString = JsonSerializer.Serialize(keyValue.Value, _jsonOptions);
+            Set(keyValue.Key, jsonString, options);
+        }
+    }
+
+    /// <summary>Serializes all the values in each KeyValue using the T type and save each into the cache</summary>
+    /// <param name="manyEntries">List of KeyValuePairs to be added to the cache, with the values being serialized.</param>
+    /// <param name="options">Optional: The cache options for the value.</param>
+    /// <param name="token">Optional. The <see cref="T:System.Threading.CancellationToken" /> used to propagate notifications that the operation should be canceled.</param>
+    /// <typeparam name="T">A class which contains the data to stored as JSON in the cache</typeparam>
+    public async Task SetManyClassAsync<T>(List<KeyValuePair<string, T>> manyEntries, DistributedCacheEntryOptions? options, CancellationToken token) where T : class, new()
+    {
+        foreach (var keyValue in manyEntries)
+        {
+            var jsonString = JsonSerializer.Serialize(keyValue.Value, _jsonOptions);
+            await SetAsync(keyValue.Key, jsonString, options, token);
+        }
     }
 }