+++ Test/TestHelpers/StubFileStoreCacheString.cs
@@ -50,6 +50,31 @@ public class StubFileStoreCacheString : IDistributedFileStoreCacheString
         return Task.CompletedTask;
     }
 
+    /// <summary>Sets many entries via a list of KeyValues</summary>
+    /// <param name="manyEntries">List of KeyValuePairs to be added to the cache.</param>
+    /// <param name="options">Optional: The cache options for the value.</param>
+    public void SetMany(List<KeyValuePair<string, string>> manyEntries, DistributedCacheEntryOptions? options = null)
+    {
+        foreach (var keyValuePair in manyEntries)
+        {
+            _cache[keyValuePair.Key] = keyValuePair.Value;
+        }
+    }
+
+    /// <summary>Sets many entries via a list of KeyValues</summary>
+    /// <param name="manyEntries">List of KeyValuePairs to be added to the cache.</param>
+    /// <param name="options">Optional: The cache options for the value.</param>
+    /// <param name="token">Optional. The <see cref="T:System.Threading.CancellationToken" /> used to propagate notifications that the operation should be canceled.</param>
+    public Task SetManyAsync(List<KeyValuePair<string, string>> manyEntries, DistributedCacheEntryOptions? options = null,
+        CancellationToken token = new CancellationToken())
+    {
+        foreach (var keyValuePair in manyEntries)
+        {
+            _cache[keyValuePair.Key] = keyValuePair.Value;
+        }
+        return Task.CompletedTask;
+    }
+
     /// <summary>Removes the value with the given key.</summary>
     /// <param name="key">A string identifying the requested value.</param>
     public void Remove(string key)
@@ -68,12 +93,22 @@ public class StubFileStoreCacheString : IDistributedFileStoreCacheString
     }
 
     /// <summary>
-    /// This clears all the key/value pairs from the json cache file
+    /// This clears all the key/value pairs from the json cache file, with option to add entries after the cache is cleared.
     /// </summary>
-    public void ClearAll()
+    /// <param name="manyEntries">Optional: After of the clearing the cache these KeyValues will written into the cache</param>
+    /// <param name="entryOptions">Optional: If there are entries to add to the cache, this will set the timeout time.</param>
+    public void ClearAll(List<KeyValuePair<string, string>>? manyEntries = null, DistributedCacheEntryOptions? entryOptions = null)
     {
         _cache = new Dictionary<string, string>();
+        if (manyEntries == null)
+            return;
+
+        foreach (var keyValuePair in manyEntries)
+        {
+            _cache[keyValuePair.Key] = keyValuePair.Value;
+        }
     }
+
 
     /// <summary>
     /// This return all the cached values as a dictionary