+++ ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 1.1.0
+
+- New Feature: SetMany / SetManyAsync now allows you to many entries in one go. This can add many entries as quickly as a single Set call.
+- New Feature: SetManyClass / SetManyClassAsync now allows you to many entries in one go. This can add many entries as quickly as a single SetClass call.
+- New Feature: ClearAll now has an option to add a list of key/value pairs into the json cache file after the cache is cleared
+
 ## 1.0.0
 
 - First release